# docs/OPENAPI.yaml
openapi: 3.1.0
info:
  title: Vyaya API
  version: 0.1.0
  description: >-
    Token-waste auditor for LLM APIs. Covers the observe-only proxy
    (OpenAI-compatible passthrough), the web app's backend-for-frontend
    routes (session cookie auth), and the worker health endpoint. Dev-only
    mock services (mock-openai, mock-deskid) are documented in docs/API.md.
  license:
    name: Proprietary
servers:
  - url: http://localhost:8787
    description: Proxy (dev)
  - url: http://localhost:3000
    description: Web BFF (dev)
  - url: http://localhost:8790
    description: Worker health (dev)
...
paths:
...
  /v1/chat/completions:
    post:
      tags: [proxy]
      summary: Proxy a chat completion
      description: >-
        Byte-faithful passthrough to the configured upstream (OpenAI or a
        KubeMind router), streamed end-to-end. The proxy logs metadata and
        cost after the client has the bytes. SSE streaming is supported;
        include stream_options.include_usage so usage comes from the
        provider instead of an estimate.
      operationId: proxyChatCompletions
      parameters:
        - $ref: "#/components/parameters/XVyayaRequestId"
        - $ref: "#/components/parameters/XVyayaSession"
        - $ref: "#/components/parameters/XVyayaTag"
        - $ref: "#/components/parameters/XVyayaRetryAttempt"
        - $ref: "#/components/parameters/XVyayaRetryOf"
        - $ref: "#/components/parameters/XVyayaConsumed"
      security:
        - vyayaApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatCompletionRequest"
      responses:
        "200":
          description: Upstream response, buffered or SSE.
          headers:
            x-vyaya-request-id:
              schema:
                type: string
              description: Assigned or echoed correlation id.
          content:
            application/json:
              schema:
                type: object
                description: Standard OpenAI chat completion object.
            text/event-stream:
              schema:
                type: string
                description: OpenAI SSE chunk stream.
        "401":
          $ref: "#/components/responses/ProxyUnauthorized"
        "429":
          $ref: "#/components/responses/ProxyRateLimited"
        "502":
          $ref: "#/components/responses/ProxyUpstreamUnavailable"
        "503":
          $ref: "#/components/responses/ProxyAuthUnavailable"
  /v1/embeddings:
    post:
      tags: [proxy]
      summary: Proxy an embeddings call
      description: Passthrough for OpenAI embeddings. Input tokens only.
      operationId: proxyEmbeddings
      parameters:
        - $ref: "#/components/parameters/XVyayaRequestId"
        - $ref: "#/components/parameters/XVyayaTag"
      security:
        - vyayaApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [model, input]
              properties:
                model:
                  type: string
                  example: text-embedding-3-small
                input:
                  oneOf:
                    - type: string
                    - type: array
                      items:
                        type: string
      responses:
        "200":
          description: Upstream embeddings response.
          headers:
            x-vyaya-request-id:
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
        "401":
          $ref: "#/components/responses/ProxyUnauthorized"
        "429":
          $ref: "#/components/responses/ProxyRateLimited"
        "502":
          $ref: "#/components/responses/ProxyUpstreamUnavailable"
        "503":
          $ref: "#/components/responses/ProxyAuthUnavailable"
  /api/auth/session:
    get:
      tags: [auth]
      summary: Who am I
      operationId: getSession
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Current session.
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    type: object
                    properties:
                      sub: { type: string }
                      email: { type: string, format: email }
                      role: { $ref: "#/components/schemas/Role" }
                      workspaceId: { type: string, format: uuid }
        "401":
          $ref: "#/components/responses/WebUnauthorized"
...
components:
  securitySchemes:
    vyayaApiKey:
      type: apiKey
      in: header
      name: X-Vyaya-Key
      description: Workspace API key (vy_live_ + 64 hex chars).
    sessionCookie:
      type: apiKey
      in: cookie
      name: vyaya_session
      description: HMAC-signed session cookie set by /auth/callback.
  parameters:
    XVyayaRequestId:
      name: X-Vyaya-Request-Id
      in: header
      required: false
      schema:
        type: string
        pattern: "^[A-Za-z0-9][A-Za-z0-9_.:-]{0,127}$"
      description: Client correlation id. Malformed values are ignored.
    XVyayaSession:
      name: X-Vyaya-Session
      in: header
      required: false
      schema: { type: string, maxLength: 256 }
      description: Session id; groups turns for context_amnesia.
    XVyayaTag:
      name: X-Vyaya-Tag
      in: header
      required: false
      schema: { type: string, maxLength: 128 }
      description: Feature tag; validated against the workspace allowlist.
    XVyayaRetryAttempt:
      name: X-Vyaya-Retry-Attempt
      in: header
      required: false
      schema: { type: integer, minimum: 0, maximum: 10000, default: 0 }
      description: Retry index. 0 = first attempt.
    XVyayaRetryOf:
      name: X-Vyaya-Retry-Of
      in: header
      required: false
      schema: { type: string, maxLength: 128 }
      description: Request id of the first attempt in a retry chain.
    XVyayaConsumed:
      name: X-Vyaya-Consumed
      in: header
      required: false
      schema: { type: string, enum: ["true", "false", "0", "1", "no", "yes"] }
      description: Downstream-consumption signal for ghost_output.
  responses:
    ProxyUnauthorized:
      description: Missing, invalid, or revoked API key.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProxyError"
          examples:
            missing:
              value:
                error:
                  message: missing X-Vyaya-Key header
                  type: invalid_request_error
                  param: null
                  code: missing_api_key
            invalid:
              value:
                error:
                  message: invalid or revoked API key
                  type: invalid_request_error
                  param: null
                  code: invalid_api_key
    ProxyRateLimited:
      description: Per-key sliding window exceeded.
      headers:
        Retry-After:
          schema: { type: integer }
          description: Seconds until the oldest in-window request ages out.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProxyError"
    ProxyAuthUnavailable:
      description: Auth backend down and key not cached. Fail-closed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProxyError"
    ProxyUpstreamUnavailable:
      description: The upstream fetch itself failed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProxyError"
    WebUnauthorized:
      description: No valid session cookie.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/WebError"
          example:
            error: not signed in
...
  schemas:
    ProxyError:
      type: object
      properties:
        error:
          type: object
          properties:
            message: { type: string }
            type:
              type: string
              enum: [invalid_request_error, rate_limit_error, server_error]
            param: { type: "null" }
            code:
              type: string
              nullable: true
              enum: [missing_api_key, invalid_api_key, auth_unavailable, rate_limit_exceeded, upstream_unavailable, null]
    WebError:
      type: object
      properties:
        error: { type: string }
    Role:
      type: string
      enum: [admin, operator, viewer]
...
    ChatCompletionRequest:
      type: object
      required: [model, messages]
      properties:
        model: { type: string, example: gpt-4o-mini }
        messages:
          type: array
          items:
            type: object
            properties:
              role: { type: string, enum: [system, user, assistant, tool] }
              content: { type: string }
        max_tokens: { type: integer }
        stream: { type: boolean }
        stream_options:
          type: object
          properties:
            include_usage: { type: boolean }
        response_format:
          type: object
          description: json_object or json_schema; the proxy validates the response against it.
...